# openapi.json
{
  "components": {
    "schemas": {
...
      "AliasCreate": {
        "properties": {
          "alias": {
            "title": "Alias",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source",
          "alias"
        ],
        "title": "AliasCreate",
        "type": "object"
      },
      "AliasRead": {
        "properties": {
          "alias": {
            "title": "Alias",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source",
          "alias",
          "created_at"
        ],
        "title": "AliasRead",
        "type": "object"
      },
      "AliasResolveRead": {
        "properties": {
          "alias": {
            "title": "Alias",
            "type": "string"
          },
          "cid": {
            "title": "Cid",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "cid",
          "source",
          "alias"
        ],
        "title": "AliasResolveRead",
        "type": "object"
      },
...
      "ArtifactRead": {
        "properties": {
          "aliases": {
            "items": {
              "$ref": "#/components/schemas/AliasRead"
            },
            "title": "Aliases",
            "type": "array"
          },
          "cid": {
            "title": "Cid",
            "type": "string"
          },
          "locator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locator"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "cid",
          "sha256",
          "locator",
          "aliases"
        ],
        "title": "ArtifactRead",
        "type": "object"
      },
      "ArtifactUpdate": {
        "properties": {
          "locator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locator"
          }
        },
        "required": [
          "locator"
        ],
        "title": "ArtifactUpdate",
        "type": "object"
      },
...
      "Body_artifacts_upload_file": {
        "properties": {
          "alias": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alias"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "source": {
            "default": "upload",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_artifacts_upload_file",
        "type": "object"
      },
      "CommitEntryRead": {
        "properties": {
          "cid": {
            "title": "Cid",
            "type": "string"
          },
          "entry_hash": {
            "title": "Entry Hash",
            "type": "string"
          },
          "prev_hash": {
            "title": "Prev Hash",
            "type": "string"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          }
        },
        "required": [
          "seq",
          "cid",
          "recorded_at",
          "prev_hash",
          "entry_hash"
        ],
        "title": "CommitEntryRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifyRead": {
        "properties": {
          "cid": {
            "title": "Cid",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "property_violated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Violated"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "cid",
          "valid"
        ],
        "title": "VerifyRead",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "aits",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/artifacts/": {
      "post": {
        "operationId": "artifacts_upload",
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "contentMediaType": "application/octet-stream",
                "title": "Body",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/artifacts/by-digest/{sha256_hex}/": {
      "get": {
        "description": "Look up an artifact by the sha256 of its raw bytes -- the comparison\npath for a client holding a file.",
        "operationId": "artifacts_get_by_digest",
        "parameters": [
          {
            "in": "path",
            "name": "sha256_hex",
            "required": true,
            "schema": {
              "title": "Sha256 Hex",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get By Digest",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/artifacts/resolve/": {
      "get": {
        "operationId": "artifacts_resolve",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": true,
            "schema": {
              "title": "Source",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "alias",
            "required": true,
            "schema": {
              "title": "Alias",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AliasResolveRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/artifacts/upload/": {
      "post": {
        "description": "Convenience: hash an uploaded file and register an alias in one call.\n\nReuses the same content-addressing as ``POST /artifacts/`` and the same\nalias upsert/conflict semantics as ``POST /artifacts/{cid_hex}/aliases/``\n(409 on a clashing pair). ``alias`` defaults to the uploaded filename, so a\nbare file upload is captured as ``(source=\"upload\", alias=<filename>)`` and\nis immediately resolvable.",
        "operationId": "artifacts_upload_file",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_artifacts_upload_file"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload File",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/artifacts/{cid_hex}/": {
      "get": {
        "operationId": "artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "cid_hex",
            "required": true,
            "schema": {
              "title": "Cid Hex",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get",
        "tags": [
          "artifacts"
        ]
      },
      "patch": {
        "operationId": "artifacts_set_locator",
        "parameters": [
          {
            "in": "path",
            "name": "cid_hex",
            "required": true,
            "schema": {
              "title": "Cid Hex",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Locator",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/artifacts/{cid_hex}/aliases/": {
      "post": {
        "operationId": "artifacts_add_alias",
        "parameters": [
          {
            "in": "path",
            "name": "cid_hex",
            "required": true,
            "schema": {
              "title": "Cid Hex",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AliasCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AliasRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Alias",
        "tags": [
          "artifacts"
        ]
      }
    },
...
    "/commits/": {
      "get": {
        "operationId": "audit_commits",
        "parameters": [
          {
            "in": "query",
            "name": "cid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cid"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CommitEntryRead"
                  },
                  "title": "Response Audit Commits",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commits",
        "tags": [
          "audit"
        ]
      }
    },
...
    "/verify/{cid_hex}/": {
      "get": {
        "operationId": "audit_verify",
        "parameters": [
          {
            "in": "path",
            "name": "cid_hex",
            "required": true,
            "schema": {
              "title": "Cid Hex",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify",
        "tags": [
          "audit"
        ]
      }
    }
  }
}